--- a/be/src/common/config.h
+++ b/be/src/common/config.h
@@ -1013,9 +1013,11 @@ DECLARE_mInt32(s3_write_buffer_size);
 // can at most buffer 50MB data. And the num of multi part upload task is
 // s3_write_buffer_whole_size / s3_write_buffer_size
 DECLARE_mInt32(s3_write_buffer_whole_size);
-
 //enable shrink memory
 DECLARE_Bool(enable_shrink_memory);
+// enable cache for high concurrent point query work load
+DECLARE_mInt32(schema_cache_cacity);
+DECLARE_mInt32(schema_cache_sweep_time_sec);
 
 // enable binlog
 DECLARE_Bool(enable_feature_binlog);