--- a/be/src/olap/rowset/beta_rowset_reader.cpp
+++ b/be/src/olap/rowset/beta_rowset_reader.cpp
@@ -17,7 +17,6 @@
 
 #include "beta_rowset_reader.h"
 
-#include <gen_cpp/PaloInternalService_types.h>
 #include <stddef.h>
 
 #include <algorithm>