--- a/be/src/olap/rowset/rowset_reader_context.h
+++ b/be/src/olap/rowset/rowset_reader_context.h
@@ -79,6 +79,7 @@ struct RowsetReaderContext {
     bool is_vertical_compaction = false;
     bool is_key_column_group = false;
     const std::set<int32_t>* output_columns = nullptr;
+    RowsetId rowset_id;
 };
 
 } // namespace doris