--- a/be/src/olap/rowset/segment_v2/column_reader.h
+++ b/be/src/olap/rowset/segment_v2/column_reader.h
@@ -34,7 +34,6 @@
 #include "io/fs/file_reader_writer_fwd.h"
 #include "io/io_common.h"
 #include "olap/olap_common.h"
-#include "olap/rowset/segment_v2/bloom_filter.h"
 #include "olap/rowset/segment_v2/common.h"
 #include "olap/rowset/segment_v2/ordinal_page_index.h" // for OrdinalPageIndexIterator
 #include "olap/rowset/segment_v2/page_handle.h"        // for PageHandle
@@ -308,8 +307,6 @@ public:
     }
 
     virtual bool is_all_dict_encoding() const { return false; }
-
-    virtual void clear() {};
 
 protected:
     ColumnIteratorOptions _opts;
@@ -586,8 +583,6 @@ public:
 
     ordinal_t get_current_ordinal() const override { return _current_rowid; }
 
-    void clear() override { _current_rowid = 0; }
-
 private:
     rowid_t _current_rowid = 0;
     int32_t _tablet_id = 0;
@@ -639,8 +634,6 @@ public:
 
     static void insert_default_data(const TypeInfo* type_info, size_t type_size, void* mem_value,
                                     vectorized::MutableColumnPtr& dst, size_t n);
-
-    void clear() override { _current_rowid = 0; }
 
 private:
     void _insert_many_default(vectorized::MutableColumnPtr& dst, size_t n);