… be/src/olap/CMakeLists.txt
… be/src/olap/schema.h
… be/src/olap/schema_cache.cpp
… be/src/olap/schema_cache.h
… be/src/olap/storage_engine.cpp
… be/src/olap/tablet.cpp
… be/src/olap/task/index_builder.cpp
… be/src/runtime/exec_env_init.cpp
… be/src/service/internal_service.cpp
… be/src/vec/exec/scan/new_olap_scanner.cpp
… be/src/vec/olap/vgeneric_iterators.cpp
… be/src/vec/olap/vgeneric_iterators.h
… be/test/olap/delta_writer_test.cpp
… be/test/vec/exec/vgeneric_iterators_test.cpp
… fe/fe-core/src/main/java/org/apache/doris/catalog/OlapTable.java
… fe/fe-core/src/main/java/org/apache/doris/planner/OlapScanNode.java
… gensrc/thrift/PlanNodes.thrift