@hello-stephen 2023-05-26
TeamCity pipeline, clickbench performance test result:
 the sum of best hot time: 33.94 seconds
 stream load tsv:          442 seconds loaded 74807831229 Bytes, about 161 MB/s
 stream load json:         23 seconds loaded 2358488459 Bytes, about 97 MB/s
 stream load orc:          58 seconds loaded 1101869774 Bytes, about 18 MB/s
 stream load parquet:          30 seconds loaded 861443392 Bytes, about 27 MB/s
 insert into select:          25.2 seconds inserted 10000000 Rows, about 396K ops/s
 https://doris-community-test-1308700295.cos.ap-hongkong.myqcloud.com/tmp/20230526175334_clickbench_pr_151644.html