@hello-stephen 2023-05-29
TeamCity pipeline, clickbench performance test result:
 the sum of best hot time: 35.33 seconds
 stream load tsv:          458 seconds loaded 74807831229 Bytes, about 155 MB/s
 stream load json:         22 seconds loaded 2358488459 Bytes, about 102 MB/s
 stream load orc:          64 seconds loaded 1101869774 Bytes, about 16 MB/s
 stream load parquet:          32 seconds loaded 861443392 Bytes, about 25 MB/s
 insert into select:          25.6 seconds inserted 10000000 Rows, about 390K ops/s
 https://doris-community-test-1308700295.cos.ap-hongkong.myqcloud.com/tmp/20230529061806_clickbench_pr_152567.html